requirements.txt

line-bot-sdk
Flask
Flask-SocketIO
pymongo
uwsgi
google-api-python-client
google-auth-httplib2
google-auth-oauthlib